FacialRecognition/requirements.txt
face-recognition==1.3.0
dlib==19.24.6
numpy>=1.26.0,<2.1.0
opencv-python==4.11.0.86
Pillow==11.1.0
tk==0.1.0
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
insightface==0.7.3
onnxruntime==1.20.1
tqdm==4.67.1
scikit_image==0.25.2
albumentations==2.0.5
onnx==1.17.0
scikit_learn==1.6.1
coloredlogs==15.0.1
sympy==1.13.3
opencv_python_headless==4.11.0.86
stringzilla==3.12.2
simsimd==6.2.1
humanfriendly==10.0
tifffile==2025.2.18
networkx==3.4.2
imageio==2.37.0
mpmath==1.3.0